Don't look up the current user in the db every time. Pass the current user to permissions service.
