+++ src/main/java/kc/ebenezer/dto/UserDto.java
@@ -1,12 +1,16 @@
 package kc.ebenezer.dto;
 
 import org.codehaus.jackson.annotate.JsonAutoDetect;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
 @JsonAutoDetect
-public class UserDto extends DtoObject {
+public class UserDto extends DtoObject implements UserDetails {
     private Long id;
     private String name;
     private String givenName;
@@ -176,5 +180,40 @@ public class UserDto extends DtoObject {
 
     public void setImageCollection(ImageCollectionDto imageCollection) {
         this.imageCollection = imageCollection;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return Collections.EMPTY_SET;
+    }
+
+    @Override
+    public String getPassword() {
+        return "secret";
+    }
+
+    @Override
+    public String getUsername() {
+        return String.valueOf(id);
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return active;
     }
 }
